requirements.txt
Flask
Flask-CORS
numpy
pandas
pymongo
python-dotenv
holidays
gunicorn